+++ src/cpp/addon-strings.h
@@ -10,10 +10,16 @@ class AddonDataStrings {
  public:
   v8::Eternal<v8::String> n;
   v8::Eternal<v8::String> readonly;
+  v8::Eternal<v8::String> RoaringBitmap32;
+
+  v8::Eternal<v8::String> OperationFailed;
 
   void initialize(v8::Isolate * isolate) {
     this->literal(isolate, this->n, "n");
     this->literal(isolate, this->readonly, "readonly");
+    this->literal(isolate, this->RoaringBitmap32, "RoaringBitmap32");
+
+    this->literal(isolate, this->OperationFailed, "Operation failed");
   }
 
  private:
